Fix mock script engine to return the value of the raw script, instead of the compiled script object
